1. Hacked a bit to allow cc and bcc to be sent upon creation of request. Only supported by rt-mailgate so far.
2. If somebody that don't have access to reply tries to reply, it is
set as a comment. My original idea was also to bounce or mail back a
warning about this, but that also struck some requestors that was
operating with several emails - and I don't want to send bounces to
the requestors, so it's commented out as of now.

3. A template correspondence-official is used for the correspondence
that goes to the requestor. This way it's possible to differ the
template to the requestor(s) from the internal templates, and also
avoid sending the headers to the requestors.

4. Changed an error message for debugging purposes.

5. Changed the kill-sub to not kill anymore. Read the comments why.