Ok, i think i significantly cut down the complexity of the actions we've got so far. I nuked the old notify rules, as they were superceeded by my more capable SendEmail.pm There are still a few things I'm not happy with... like, the way that scrips, scripscopes and templates are related. it needs some thought and cleanup. but i think it'll work just fine for now.
I redid the templates and scrips to fit with the revised actions and added
some more docs to SendEmail.pm

the generic RT::Action should deal with templates, not SendEmail, I think,
but I'm leaving that abstraction for another day.

Oh, the Argument from the scrip is now passed into the template as $T::Argument.