Cleaned up a couple minor things (like precedence, again) and changed "FixSubject" to "SetSubjectTag"
Fixed the memory leak in Transaction.pm (it wasn't letting go of Scrip objects)
The fix was a bit of a "large mallet on a small nail" in that I'm now _explicitly_ destroying ScripObjects two different ways and don't quite understand why I have to do it. It does, however, work.