Bumped the required version of SearchBuilder in testdeps, now that it's propagated throughout CPAN
Replaced initdb.mysql with ivan's new version.

Tweaked ivan's initdb.mysql to be a little friendlier, create the schema
(it was missing a FILEHANDLE in a print statement and deal
better with omitted passwords.

Updated the readme some more.

Added experimental ChangeLog generation to the make dist process

Bumped the version number to 1.3.18 for release as alpha1 for RT2 today.