Refactoring how ticket comments and correspondence are submitted.
Comment and Correspond now share a subroutine, rather than a bunch of identical code.

RT::Transaction::Create now holds onto a bunch more Scrip data and metadata

Scrips being applied now expose the set of scrips that actually get prepared.

It's possible to create a transaction that prepares, but does not fire scrips. 

The API for accessing this has not yet been stabilized

git-svn-id: svn+ssh://svn.bestpractical.com/svn/bps-public/rt/branches/rt-3.1@288 e417ac7c-1bcc-0310-8ffa-8f5827389a85