RT-Ticket: 6774 RT-Status: resolved RT-Update: correspond
Actually set password when creating a new user, and clean up the error reporting
from password setting.  (And if you don't manage to set a password (either by leaving
it blank or by entering an invalid password) when creating a user, add a notice that
the user won't be able to log in.)  Also, indentation cleanup.

Patch by Ruslan Zakirov.

git-svn-id: svn+ssh://svn.bestpractical.com/svn/bps-public/rt/branches/3.4-RELEASE@3389 e417ac7c-1bcc-0310-8ffa-8f5827389a85