Performance Improvement when Sending Email using sendmailpipe -
MIME::Entity would bog down in certain cases because of it's use of IO::Scalar during stringification.  MIME::Entity will be switching to IO::ScalarArray, which will help... but RT was causing it to store into a temporary string anyway, which was silly.

This change has MIME::Entity write directly to the pipe, which is a lot more efficient.  Seems to cut out ~33% of user time.  (Because we don't need to have a temporary IO::Scalar thingy around.)  Also will reduce peak memory usage.

git-svn-id: svn+ssh://svn.bestpractical.com/svn/bps-public/rt/branches/3.4-RELEASE@3892 e417ac7c-1bcc-0310-8ffa-8f5827389a85