RT-Ticket: 7136 RT-Status: resolved RT-Update: correspond
Stuart Knight reports:

As part of the "initdb" processing, the scripts went through and created a new database user, in my case called RT3.

When it came time to create the tables, the script was still logged on as the dba user "system", so all of tables/sequences were created under "system"'s schema.

I followed through the rt-setup-database script, and spotted that there was a database disconnect, followed by an immediate reconnect, as the same user. (in the case of Oracle this still being the "dba" account)

Putting an extra validation check in here for Oracle, and then connecting as the intended database user fixed up the issue.

git-svn-id: svn+ssh://svn.bestpractical.com/svn/bps-public/rt/branches/CHALDEA-EXPERIMENTAL@4192 e417ac7c-1bcc-0310-8ffa-8f5827389a85