MySQL >= 4.1 specific scheme switch to BLOBs where it's required set ascii as charset in some places define default charset for tables require DBIx::SB 1.44 (not yet released) as it has better   support for DatabaseVersion method in rt-setup-database scan for versions files, for example 'schema.mysql-4.1' apply only version files that are smalle then the version of the RDBMS system,    so schema.mysql-5.0 wouldn't be applied if user's using mysql 4.1 apply closest one version file
All this changes were done in order to address issue that could hit users who are using
RT with mysql 4.1 with default charset UTF-8 or other charsets that has illegal byte
sequences. RT was using TEXT fields for binary data (for example for attachments content),
but in MySQL 4.1 and greater charsets handling was improved and mysql and requirements
to the TEXT (CHAR or VARCHAR) fields become more strict. So we have to convert all
fields that should handle arbitrary binary data into BLOBs (BINARY or VARBINARY).

git-svn-id: svn+ssh://svn.bestpractical.com/svn/bps-public/rt/branches/3.5-EXPERIMENTAL-MYSQL-UPDATES@5149 e417ac7c-1bcc-0310-8ffa-8f5827389a85