RT-Ticket: 8781 RT-Status: resolved RT-Update: correspond
I've encountered a problem with sender identification in the mailgate. If
a user sends an e-mail where 'Reply-To' is present but empty (i.e there is
'Reply-To: ' in the e-mail headers), RT does not associate the e-mail
with a RT user even if the user exists in the system and the sender
address is specified in the 'Sender' or 'From' fields.
 - Patch from ondrasej@centrum.cz

git-svn-id: svn+ssh://svn.bestpractical.com/svn/bps-public/rt/branches/3.6-RELEASE@10129 e417ac7c-1bcc-0310-8ffa-8f5827389a85