Fix display problem with non ascii headers in forwarded emails.
We forward OriginalContent, which uses OriginalEncoding, but we forward Headers
for this content as recorded by RT i.e. always using UTF-8. If content is text
and uses different encoding than UTF-8, display in MUA is broken.
So this patch encode headers as RFC 2047 to make sure MUA can read them.

git-svn-id: svn+ssh://svn.bestpractical.com/svn/bps-public/rt/3.8/trunk@19580 e417ac7c-1bcc-0310-8ffa-8f5827389a85