replace OriginalHeaders with EncodedHeaders
There is no such thing as OriginalHeaders, if a field in the head
contains not-ASCII characters then you have to use Q or B encodings.
These encodings define character set of the content and can have
multiple chunks with mixed character sets. We don't store all these
variations in charsets across fields, so we can not restore original
headers.

As far as I know OriginalEncoding can not be base64 or quoted printable,
so some parts of the code has never worked.

For purpose of "Download with headers" feature it's better to have
EncodedHeaders method that just encodes headers into a character set.