Move all JS for hierarchical CFs onto derivative field; remove DerivativeCFs method
Oracle does not allow for SQL queries which attemt to match on the
'Content' field of attrbiutes, as it is a CLOB column -- additionally,
searches on it are not likely to be well-indexed.  Because of this, we
move all code involving hierarchical CFs onto display of the
derivative field, instead of being split between base and derivative
field; this saves us the need to list all derivative CFs.

As part of this, remove the (non-functioning under Oracle) and
now-unused DerivativeCFs method on CustomField.