Stop sorting Callbacks directories
Originally this was added because all callbacks lived in
local/html/Callbacks and we needed to impose some order other than "how
readdir returned a list of directories".

Now with plugins, each Callbacks directory lives in
local/plugins/RT-Extension-Foo/html/Callbacks and we have a user defined
sort order (the order of the @Plugins config option) which we should use
instead.