RT::Init() needs to run before use RT::Tickets;
Because RT::Init pushes Plugin Paths onto @INC at run time, if we
use RT::Tickets; it will never see a Tickets_Local.pm in a plugin.
The best way to fix this is with a BEGIN block around RT::Init, but this
will really screw up rt-crontool's ability to pull LogLevel from the
command line, so we're doing the 'switch to require' approach instead.