ProcessACLChanges now expects values from a series of checkboxes
This breaks the old SelectRights element, but makes it work for the new
rights editor.

It now expects form inputs with names like
SetRights-PrincipalId-ObjType-ObjId instead of Grant/RevokeRight.  Each
input should be an array listing the rights the principal should have,
and ProcessACLChanges will modify the current rights to match.
Additionally, the previously unused CheckACL input listing
PrincipalId-ObjType-ObjId is now used to catch cases when all the rights
are removed from a principal and no SetRights input is submitted.