Tidy up and re-layout the new branding/theme page
Hopefully these changes make it more user-friendly and easier to see
what's going on.  There are still a few usability changes we need to
make, including a logo delete button and handling GIFs more gracefully.

We should also improve the selectors we use to account for links not
changing with text color.