Remove tabbing from ticket update page
This is in preparation for making the squelch on this page be for the
current transaction only, not a durasive squelch.  Between having to
store "people to squelch on the upcoming unborn transaction" or alter
the UI to have that information on the same page, the latter seemed
more correct.