Clean up and standardize the wording of right descriptions
Descriptions should lead with a verb and avoid phrases like "this queue"
or "this group" since those don't make sense when applied in a global
context.  They should be concise but helpful.  Rights which are
logically grouped together should be phrased similarly.

Begin a glossary of terms, starting with the standard right name
prefixes.