Limit articles to classes applied to a queue
Now that we have ObjectClasses and can declare a class to be Global or
applied to one or more Queues, make it possible to limit Articles
returned by searches to those Classes which are relevant.  A lot of this
logic is similar in concept to what we are doing with Custom Fields.