Updated RedHat layout:
I installed RT with configure option --enable-layout=RH.

Subsequent installs of RT-FM and RT-IR caused the installer to issue a
message that it could not find the RT.pm package because it is located
under ../rt and the installer is looking for ../rt3.

The RH layout should use rt3, not rt.