Redirect to the desired page after logging in a user
This prevents back button attacks in the form of resubmitting form data
after the user has logged out of the browser (but not closed it).  See
also rt3 #15804.

For non-homepage hits, the browser now also gets redirected to
/NoAuth/Login.html to get a login form.

We use the session to store the next page URL (referenced by a hash),
similar to how action results work.