Revert "Import history folding from a client project, somewhat retooled"
This reverts commit 656ca37bd05c055bdc9c66a97450786db33e8666.

Conflicts:

	etc/RT_Config.pm.in

The original commit broke the HTML of the page by leaving <div>s
unclosed.  It also failed to take into account the message formatting
options PlainTextPre and PlainTextMono.

We want the history folding functionality, but it needs to work
correctly first.  I'm moving it out to a branch for now where it can get
fixed up.