Enforce unique group names for user defined groups
Existing non-unique groups will continue to function, but it is now not
possible through the UI or API to create another non-unique group or
rename an existing group to an already taken name.

The Pg and SQLite schemas have required this for 7-8 years, but the
MySQL and Oracle schemas never have.  A tool to report non-unique group
names will appear in the near future.