Switch the default scrip templates to HTML for new installs
This also switches the default for testing.  As an escape hatch,
RT::Test now sports a text_templates flag.

Tests which failed non-obviously with HTML templates enabled were copied
to a t/mail/*-plaintext.t filenames and flipped back to testing
plaintext.  The original files were fixed to expect the New World Order
of HTML.  This way both the new HTML templates and the old plaintext
templates are tested.