Add a create form to the admin groups page
This refactors the group modify page to split the form out into a
reusable component EditGroup.  Modify.html still does all the work and
the form is submitted there.  Now we don't need the annoying
Select/Create submenus in the admin menu.

The layout on the admin groups page needs some love, but the
functionality is present.