Revert "Kill the superfluous Settings submenu to decrease the menu levels" Revert "Simplify the Logged in user menu in the web2 theme"
This reverts commit e5d457cccbb0187623a71cbe17e5e29051ac1ba5.
This reverts commit c4dba6b4fc9d3a3dd89db93b3642fb8e42174b4c.

Things like "Saved searches" are the settings for those saved searches,
and are potentially more confusing without the submenu.  The submenu
does correctly group all of the items underneath it.  I'm still not
entirely happy with this deeper menu structure, so suggestions for
improving it are welcome.  Simply moving the Settings menu to the Tools
menu doesn't seem right either.