Clean up the menuing code to remove some cruft and unnecessary hacks
Kill the link_item attribute.  We can figure that out based on other
attributes.  Actually use the raw_html attribute.  escape_title stays
because it is still useful as a smaller hammer than raw_html.  Respect
the class attribute.  Cleanup and correct the doc.

Link Tickets -> Current Search to the search results page, which is
handy.

Ideally we wouldn't wrap menu items without a path in an <a>, but we
still need to for the sake of the menu styles.  If we wrap unlinked
titles in span.menu-item's, we can style on .menu-item instead of a.