Refresh the titlebox title tabs a little for aileron
The rollup/down arrows are now sprites so we can use a different color
in aileron but preserve the original color for older themes.

The top padding on .titlebox-content is a little much when there isn't a
right side links bar.  But without it most content runs into the link
bar.

Still needs testing in IE and undoubtedly fixes.