Allow ignoring of current columns when upgrading through old versions
RT 3.9.5 added three columns to the ACL table; if the user is
upgrading through any versions before that which attempt to add rows
to the ACL table (3.8.2, 3.9.1), they will fail.  Work around this by
temporarily removing the columns from RT's internal schema if the an
upgrade predates the columns being added.  To determine when key
columns were added, look for a file named 'backcompat' in the
appropriate upgrade directory.