Let the admin know if they try to add rights for an invalid principal
Show a warning message and disable the checkboxes.  Note that for users
you must enter a Name if you ignore the autocompleter.

If the user types a custom value ignoring the autocompleter, than an
ajax check is done to the autocomplete helper to verify that the value
is valid.  If a value is picked from the autocompleter, then the extra
ajax check is skipped.