Prevent hangs during testing by removing a signals race condition
Under high load, such as heavily concurrent testing, there is no
guarantee that the parent process of the fork will have time to set up
the signal handler before the shild finishes initialization.  Thus,
receiving a USR1 signal causes it to die, leaving a zombie child
process and the test wedged.

Set up the signal handler before we fork, which guarantees that the
child will not be able to force this race condition.  It leaves forked
server processes with a USR1 signal handler, but there is little harm
to this.