Avoid sleeping for 15s if the server is already ready
It is possible for the server to have sent us the ready signal before
the parent hits the sleep(); in this case, the parent would wait,
uninterrupted by SIGUSR1, for 15s, despite the server being ready.

Hence, only sleep if we have not already received the signal.  There
is still a minute race condition between the "unless" check and the
"sleep," but this reduces the window considerably, and the side
effects to the race condition are merely a slower test.