Improve custom user logo caching with the same MD5 trick JS/CSS uses
Bad caching here is more painful than with the JS/CSS because it's an
easy, user-changeable item that looks like a bug.  It affects end users
of the system since they won't see the logo change or know to force
refresh.

The dhandler serves up the current user logo, if we have one, regardless
of the url after /NoAuth/Helpers/CustomLogo/.  Users of release
candidates who uploaded a logo won't experience a broken logo after
upgrade because their attribute doesn't have the hash key.

This commit also refactors /Admin/Global/Theme.html to display the logo
using /Elements/Logo as well to avoid code duplication.  A pleasant
side-effect is that the default BPS logo shows up as the current logo
when there is no user logo; previously it was just blank.