Replace 3384147 with a less error-prone explicit loop
An else-less 'if' inside a map produces the single value '' when the
if is false (the return value of the test).  In this context, this
would likely lead to an odd number of elements in the hashref, instead
of the pre-3384147 result of simply having a '' key.  Additionally,
perlcritic warns (not incorrectly) about the construct.

Replace the map with an explicit for loop, which makes perlcritic
happy, in addition to solving the "odd number of elements" bug above.