Allow prefixing of the cflabel/cfentry html ids in EditCustomFields
If you have two ticket creation / editing pages in the same web page,
you'll end up with multiple tds that have the same id which is invalid
and causes bad things to happen with javascript.  This provides a hook
to let you tweak the id to form2-cflabel-id on your second form.