first pass on stop calls from callbacks
This experimental code that uses Sub::StopCalls
module to jump over call to $m->callback method
second time when we're sure there is no callback.

Yes, we have a cache in the method, but code
still does things before it gets to the cache
check.

The module rewrites perl program itself (op tree)
to avoid any further calls from the current
location. It's a bit extreme, but we may consider
it with an option in the config.