Disallow SQL injection in FIELD argument to OrderBy
A number of locations blindly pass OrderBy clauses from the user into
the FIELD of OrderByCols, which provides no escaping or sanity checking.
As a short-term fix, limit the FIELD to /^[a-zA-Z0-9_]*$/ at the
RT::SearchBuilder level.

There is one place which injects logic into the ORderByCols FIELD;
namely, the "PAW sort" feature, which sorts the current user's tickets,
followed by unowned tickets, followed by all other tickets.  Allow this
codepath by adding a FUNCTION field to OrderByCols which steps around
the limit mentioned above.  This particular use is safe because no
user-specified data is used in the function which is built.

The severity of this SQL injection is limited by the ability to create a
single dangerous statement by appending to a "SELECT ... FROM ... ORDER
BY main." statement.  The lack of parens around the SELECT statement
precludes a UNION attack, and the use of DBI prepared statements
disallows the attacker from running more than one SQL statement, thus
precluding destructive queries such as "DELETE FROM" or "DROP TABLE".