Prevent FIELD- and OPERATOR- based SQL injection at the RT::SB level
Provide a whitelist of possible operators to use, and restrict field to
being /^\w*$/.  Any arguments not matching this will instead log a
critical message and insert a clause which should never match.

In the two places we attempt to be clever, and generate the SQL
"FUNCTION( alias.field ) = 42" by putting "FUNCTION( alias" as the
ALIAS, and "field )" as the FIELD -- which violates this constraint.
Provide a new FUNCTION parameter, which walks around the new
restriction, and does the split across the '.' for you.  This should
only be used if you are certain that you trust the data that the
FUNCTION argument is built up from.