Stop direct access to richtext editor files
We have a dhandler that rewrites access, but nothing that stopped you
from directly invoking one of the files under the RichText editor.
This means you could ask mason to run a .js file or a .png.
You shouldn't be able to do anything interesting with this, but it seems
wrong so we're stopping it by providing an autohandler.

However, the .js or .png is still compiled before mason consults the
autohandler.  Please don't add BEGIN blocks to your .js or .png files

We should really have static content outside the mason component root