Cherry pick back d9f6bba062815e698298f6bd04fe312647123011
Ensure we try to load RT::Interface::Web_(Overlay|Local|Vendor).pm

This was broken because RT/Interface/Web.pm declares package
HTML::Mason::Commands before the _ImportOverlays call.  Other packages
don't seem to be similarly broken, and we have a better solution in mind
for 4.2.

As an aside: the new overlay loading requires that you declare a package
in your overlay (otherwise you're in the RT::Base package).  There's no
clean way to require a file in the context of a specific package without
hardcoding the package name, so we can't just fix _ImportOverlays to
declare a package for you.

See ticket #17170 and http://lists.bestpractical.com/pipermail/rt-devel/2011-April/011446.html