Walk around acls when determining transaction type
It is entirely possible that you'll have a ticket created by an
Unprivileged user without ShowTicket.  This means that the Ticket Create
transaction will have a CurrentUser of that Unprivileged user.
$transaction->Type does not work if you don't have ShowTicket.
This means that _FindScrips will go look for TransactionBatch Scrips
with an ApplicableTransTypes of Any, not Create and you'll never have an
On Create TransactionBatch scrip run.