Actually use the tsvector column and index that we added
Failing to do this means that we fail to actually use either our
existing tokenizations or our full-text index over them, causing
postgres to re-tokenize the content of every row in the process of doing
the search.  This causes the search to be two orders of magnitude slower
than simply using LIKE.

It incidentally also skips the (not-yet-used, but stubbed in
rt-fulltext-indexer) step where we would cleverly downgrade text/html
attachments before tokenizing them.