Work around RT::Tickets->Limit not actually doing limiting
Without this, the JoinTransactions called from a RT::Tickets object
method fails to apply the ObjectType limit.  This is both incorrect, and
means that the query does not use the appropriate index on Transactions.

This happens beacause RT::Tickets->Limit does not actual limiting, but
attempts to transform its arguments to Restrictions, to TicketSQL, to
actual limits.  Somewhere along the way, the ObjectType gets silently
lost.