Disallow configuring the query column in the Sphinx table
Sphinx assumes that the third column in the table is used for issuing
queries; its name does not matter.  However, as it is an entirely new
table, there is no reason to allow configuration of the name of the
column; it simply increases complexity and gives the illusion of
control.  Hard-code it to the suggested 'query' column.