Configure the number of attachment results to return from Sphinx
The default is to only return the first 20 matches; this can be
increased by raising the 'limit' parameter, but only up to the system
max_matches value, which defaults to 1,000.  At FTS setup time, prompt
for the value to use, and default to the higher value of 10,000.  This
can still cause problems wherein the first 10,000 results from Sphinx do
no overlap with the other query restrictions (queue, ticket status,
etc), and thus RT will incorrectly return no results.