We can still use Test::Warn for this case, albeit with a different check
Test::Warn only returns the first line of the warn, unfortunately.  This
particular case is more verbose than "bad passphrase" -- it notices that
it has +no_ passphrase given, and throws an additional warning that it
can't ask for the passphrase from the user, as it is being run
non-interactively.  While we'd ideally also test for "bad passphrase",
simply testing for "can't query passphrase in batch mode" is not wrong.