Revert "Oracle needs TO_NUMBER to store clob in number column"
This reverts commit ec0ef6c832497ad76da972e4f499fcd7eb371799.

BasedOn can have multiple duplicate Attributes (#17356) which means this
SQL needs to have a LIMIT 1 (or the oracle equivalent which is too long
to fit in this commit message).
Instead, since BasedOn is a special column, we'll delete all the
database specific logic and just use perl, like the other 2 parts of the
upgrade script.