RT::Test::create_ticket(s) had a silent requirement on Queue id
When we started correctly testing keys of the hash passed in after
b65f76905af692cdfafacb509f869785ac3e2ddd, we call Ticket->Queue which
is an id, not a name.  This means anyone calling create_tickets()
without a Queue and picking up the default of Queue => General would
fail a test in an obscure way, and anyone calling create_ticket(Queue =>
'General') would fail when it expected 1.