Work with our new multipart handling
As part of our outlook handling code, we started recursing into
multipart/alternative.  However, if you were multipart/multipart or
something else, and the code doesn't express a preference for text/plain
vs text/html we had a bail out to start at the top again looking for
textual.  Instead, we'll look at the children of this multipart for
anything textual, since that seems to make more sense and leave the
"Search for something textual from the top" for later in the code path.