Use a locked file to eliminate race conditions when picking a port and db name
The previous method of picking a random port and testing if it was
unused failed under high concurrency, as there is a large gap between
when the port is checked for availability and when the server is started
to claim it.  This is particularly important because (in parallel
testing) the database name is based on the port, so failures show up
even if there is no server started in the test, as the computed database
names conflict.

Thus, provide a flock()-protected file which provides a list of ports
which we know to have claimed, as _well_ as checking the port for being
already bound.  This provides a stronger requirement for the race
condition to be met.