During user and group creation, keep submitted values on error
If the user creation fails, pull the defaults for the form from the
values submitted previously.  This is not entirely possible for
checkboxes, but this provides simple best-effort towards not losing
user-input data.

We intentionally insert $Object->Field||$Field, and not the other way
around, such that invalid values are preserved on create (to reduce
information loss) but are reset on update (to prevent the user from
assuming that they succeeded).  Since field updates are orthogonal
during update (a failure to update email doesn't prevent phone from
updating), this is not a significant difference, however.