Prevent the session cookie from being read by client-side javascript
Setting HttpOnly prevents "document.cookie" in javascript from returning
the cookie in question (other cookies from the same host are
unaffected).  Since the session-id of the request is never used by
client-side javascript directly, this prevents it from being read and
(and forwarded on) by malicious javascript.  Obviously, this is only a
paper-thin layer of fallback security, as there are many other avenues
for exploitation that malicious javascript could take.