Properly match on either UpdateType or Action
Previously, having UpdateType set simply caused it to assume
correspondence, due to precedence.  This was visible as the second time
a Ticket Update page was loaded (due to GPG failure, "Add More Files",
or whatnot), the Preview Scrips pane would act as if the transaction
were a correspondence.  Fix this by properly adding parens to the regex
target, which was improperly refactored in 25792acd.