Use CountAll for total tickets in the Mobile UI ticket list
CountAll does the count ignoring the RowsPerPage setting, where Count
will be the total records in the page.

Count and CountAll actually work the same here since a query isn't
processed yet, but having CountAll here makes it easier to understand what
the title is trying to portray (total tickets in the search regardless
of page count).