Work around a bug in perl < 5.12 where use of => enables the utf8 flag
The fat comma, on perl 5.10.x and below with the "use utf8" pragma
enabled, turns on the internal UTF-8 flag on the string it generates,
regardless of if it contains codepoints > 127.  As we erroneously use
the presence of the UTF-8 flag to determine if the string we are given
is already encoded, the this causes different results, despite that the
two strings would test as being 'eq'.

For the purposes of tests, which are the only place which use the "use
utf8" pragma, use double-quotes instead of relying on the fat comma to
quote the keys, which works around this bug.