Save on data trasferred on the login page
Sending the rich text editor javascript is an extra ~330k of javascript,
which is unnecessary on the front page, and waiting for it to transfer
leads to perceptions of slowness.  Additionally, since we currently
serve the full login page to mobile devices, this saves a significant
amount of data sent to mobile devices.