Rework the saved search tab logic to allow gramular arg overrides
/Elements/Tabs provides a default of arguments for the query string
that gets tacked onto each tab link. Instead of throwing those defaults
away when passing a QueryString. This change lets you pass in QueryArgs
instead of QueryString if necessary. QueryArgs basically 'merges' into
the default query hash, gets, serialized, and appended to the links.

Some code outside of this logic (such as RSS links) was affected by this
change in an effort to refactor the query building.