Add the ability to include existing attachments when responding to a transaction
The ticket update page (and any other place which uses
/Ticket/Elements/AddAttachments) will offer up the named attachments
from the transaction being quoted (if any).  Each selected attachment
causes an RT-Attach: id header to be added to the update.

The RT-Attach headers are inspected to actually add the attachments to
outgoing mail in RT::Action::SendEmail, as long as the creator of the
transaction with the RT-Attach headers can see the requested
attachments.  If the outgoing template has a false RT-Attach-Message
value, then RT-Attach headers are skipped just like attachments directly
on the transaction.

Any RT-Attach headers in the template are merged with the ones from the
transaction.  This makes it possible, if permissions are correct, to use
a template-level RT-Attach header to attach arbitrary content.

This feature is accessible via the Perl API by passing in the
AttachExisting option (as an arrayref) to RT::Ticket->Correspond and
RT::Ticket->Comment, or manually setting RT-Attach headers on
transaction content.