Standardize package preamble a bit
For consistency, the first lines should now be:

    package RT::Whatever;
    use strict;
    use warnings;
    # no warnings qw( redefine );  # if necessary

    # use base qw( PackageName );  # if necessary

    # other use statements

This often moved "use strict" and "use warnings" below the package line;
it is scoped to the nearest block, and hence its location relative to
the "package" statement is irrelvant.  It also moved the "use base"
line, if present, to the top of the use statements, and, when possible,
removed the "use PackageName" line if the module already had a
"use base qw( PackageName )", as the latter implies the former.