Only check the u+x bit, not the entire mode
Git does not track the excutable bit as more than one global bit; as
such, there is no way to check in a file which has mode 0754 and
guarantee that it will be checked out as such later.  With no other
information, git will re-create it 0755 -- or, if the user's umask is
0077, it will be recreated as 0700!  Explicitly only check the u+x bit,
which is all that git can reliably set.