Pass the Queue so that Articles can limit to it
Otherwise you get an ugly error in the logs and Articles refuse to limit
the Classes to the current Queue.  This means an article in any class or
a hotlist in any class (given that you can see them) would show up in
the search, since we're not correctly limiting to the Queue.
(cherry picked from commit be372d2b90b8d95af4549fe31fe0cf3908fa6c15)