Fix CollectionList links to global scripts on queue admin page
The "Scrips" format defaults to pulling from the current request path
(WebRequestPathDir), which in the case of ListGlobalScrips is
"/Admin/Queues", as that element is only called from
/Admin/Queues/Scrips.html  However, those rows should instead link to
/Admin/Global/Scrip.html

Fix this by breaking the admin formats in two -- one for global scrips,
and one for queue-specific scrips.  These CollectionLists will hence work
from any request path.

This is not an upgrade hazard because hashes set in RT_SiteConfig.pm by
way of `Set( %Hash, ...)` override only the keys that they specify.  The
addition of the GlobalScrips key in %AdminSearchResultFormat will hence
be picked up even if the local configuration has a
Set( %AdminSearchResultFormat, ... ) present that does not contain a
GlobalScrips key.