In Postgres, simply skip attachments whose tsvectors are too large
While PostgreSQL's tsvector format has a number of limitations [1], one
of them is particularly important:

     * The length of a tsvector (lexemes + positions) must be less than
       1 megabyte

This causes an error which causes the statement to fail, and thus
rt-fulltext-indexer stops there.  Detect that particular failure case,
and allow indexing to proceed, simply skipping that attachment.

The rest of the limitations, and the reasons they do not need to be
dealt with similarly, are:

     * The length of each lexeme must be less than 2K bytes

This produces a warning from Postgres, but the statement completes.

     * The number of lexemes must be less than 2^64

This limit is hit far after the 1 megabyte limitation above, and thus
essentially can never be triggered.

     * Position values in tsvector must be greater than 0 and no more
       than 16,383

If a lexeme is found later than position 16,383, it is simply capped at
being position 16,383.  This causes information loss, but as RT does not
present the position of the match in any form, this does not matter.

     * No more than 256 positions per lexeme

Only the first 256 locations are stored; later occurrences are ignored.
Again, as RT does not present the locations of the matches, this
information loss is irrelevant to the application.

[1] http://www.postgresql.org/docs/9.1/static/textsearch-limitations.html