Preserve header values which start with their own header names
Mail::Header (the parent class of MIME::Header) has the following
helpful property:

    $ perl -MMail::Header -e '$h = Mail::Header->new;
                              $h->add("Subject", "Subjec: moose");
                              print $h->as_string'
    Subject: Subjec: moose

    $ perl -MMail::Header -e '$h = Mail::Header->new;
                              $h->add("Subject", "Subject: moose");
                              print $h->as_string'
    Subject: moose

To preserve the header precisely as we had it previously, prepend the
value with a space when we re-add it.  This space is stripped before the
header is stored, but suffices to prevent the matching prefix from being
removed.