Fix quoting in GetReferencedQueues to return useful keys in the hash
Previously, GetReferencedQueues returned queue names (including the
outer quotes, with internal escaping preserved) or possibly queue ids,
exactly as they were in the TicketSQL.  The code which used this return
value had to correctly strip the outer quotes, as well as remove the
escaping.  They chose to do this with a perplexing regex which worked
more by accident than anything else, and did the unescaping improperly
or not at all.

Standardize on returning the names in a manner suitable to be passed
unmodified to $queue->Load.