Only return queues which were referenced explicitly and positively
Previously, a TicketSQL query of << Queue != 'General' >> would
incorrectly limit the custom fields in the QueryBuilder to those on the
General queue.  Since limiting to only queues referenced with the '='
operator is nearly always more correct, only return those queues.  While
this is clearly insufficient to solve the more general problem (it still
fails on << Queue = 'General' OR Queue != 'General' >>, for example), it
in general produces more correct results.