Bump the FCGI dependency, which closes FCGI's CVE-2011-2766
CGI::Fast, which RT 3.8 uses for its FastCGI interface, uses a
deprecated interface of the FCGI module.  The FCGI module, in release
0.70 (22 March 2010), introduced a bug in this interface, wherein under
mod_fastcgi (but not mod_fcgid) the environment of the very first
request to the FastCGI child was copied into all subsequent requests.
Among other things, this means that the cookies of the first request
were seen by all subsequent requests that did not themselves specify a
cookie.

FCGI version 0.74 closes this bug; bump RT's dependency accordingly.