Specify a shorter socket path when running tests under mod_fastcgi
UNIX socket paths are required to fit in the size of the sun_path field
of a sockaddr_un, which is a char[108] on modern Linux machines, or a
char[104] on OS X.  Apache will error out if the socket path for the
fastcgi server is longer than this limit.

By default, the socket path is generated by taking the FastCgiIpcDir and
appending an md5sum of the fastcgi process' path, user, and group.
Since 8f40138, the per-test tmpdir's path is longer for tests with
longer filenames.  This, combined with the 32-character md5sum, can
cause the generated socket path to be greater than the 108 or 104
character limit, if the top level rt/ directory is not close to the
filesystem root.

Remove (32 - length("socket")) = 26 characters from the socket path by
setting the socket path explicitly.  This is still possible to overflow;
a more robust fix would be to place the socket under /tmp, but this
removes some of the locality of the temporary files.  Care would also
need to be exercised in the naming of the socket, as mod_fastcgi
generates the same md5sum for all test runs, as they all have the same
fastcgi server path, user, and group.